+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/chat-utils/anthropic.test.ts
@@ -1,3 +1,5 @@
+import { normalizeConversation } from "../ModelTraceExplorer.utils";
+
 export const MOCK_ANTHROPIC_INPUT = {
   max_tokens: 1024,
   messages: [
@@ -44,3 +46,25 @@ export const MOCK_ANTHROPIC_OUTPUT = {
     service_tier: 'standard',
   },
 };
+
+describe('normalizeConversation', () => {
+
+  it('should handle anthropic input', () => {
+    expect(normalizeConversation(MOCK_ANTHROPIC_INPUT)).toEqual([
+      expect.objectContaining({
+        role: 'user',
+        content: expect.stringMatching(/describe this image/i),
+      }),
+    ]);
+  });
+
+  it('should handle anthropic output', () => {
+    expect(normalizeConversation(MOCK_ANTHROPIC_OUTPUT)).toEqual([
+      expect.objectContaining({
+        content: expect.stringMatching(/this is a close-up macro photograph of an ant/i),
+        role: 'assistant',
+      }),
+    ]);
+  });
+
+});
