--- mlflow/server/js/src/shared/web-shared/model-trace-explorer/chat-utils/dspy.test-utils.ts
@@ -1,7 +1,0 @@
-export const MOCK_DSPY_INPUT = {
-  question: 'What is the capital of France?',
-};
-
-export const MOCK_DSPY_OUTPUT = {
-  answer: 'The capital of France is Paris.',
-};
