+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/chat-utils/gemini.test.ts
@@ -1,3 +1,5 @@
+import { normalizeConversation } from "../ModelTraceExplorer.utils";
+
 export const MOCK_GEMINI_INPUT = {
   model: 'gemini-2.5-flash',
   contents: 'Explain how AI works in a few words',
@@ -77,3 +79,25 @@ export const MOCK_GEMINI_OUTPUT = {
   automatic_function_calling_history: null,
   parsed: null,
 };
+
+describe('normalizeConversation', () => {
+
+  it('should handle gemini input', () => {
+    expect(normalizeConversation(MOCK_GEMINI_INPUT)).toEqual([
+      expect.objectContaining({
+        role: 'user',
+        content: expect.stringMatching(/explain how ai works/i),
+      }),
+    ]);
+  });
+
+  it('should handle gemini output', () => {
+    expect(normalizeConversation(MOCK_GEMINI_OUTPUT)).toEqual([
+      expect.objectContaining({
+        content: expect.stringMatching(/ai learns patterns from data to make decisions/i),
+        role: 'assistant',
+      }),
+    ]);
+  });
+
+});
